requirements.txt
asgiref==3.5.2
cssselect==1.1.0
Django==4.1
django-jsignature==0.11
lxml==4.9.1
Pillow==9.2.0
pyquery==1.4.3
sqlparse==0.4.2
tzdata==2022.2
